+++ src/Plutarch/Extra/FixedDecimal.hs
@@ -2,9 +2,14 @@ module Plutarch.Extra.FixedDecimal (
   FixedDecimal (..),
   fixedNumerator,
   fixedDenominator,
+  emul,
+  ediv,
+  convertExp,
   PFixedDecimal (..),
   pfixedNumerator,
   pfixedDenominator,
+  pemul,
+  pediv,
   pconvertExp,
   pfromFixedDecimal,
   ptoFixedDecimal,
@@ -14,7 +19,7 @@ module Plutarch.Extra.FixedDecimal (
 
 import Data.Proxy (Proxy (Proxy))
 import GHC.Real (Ratio ((:%)))
-import GHC.TypeLits (KnownNat, Natural, natVal)
+import GHC.TypeLits (KnownNat, Natural, natVal, type (+), type (-))
 import Plutarch.Extra.Rational ((#%))
 import Plutarch.Lift (
   PConstantDecl (PConstantRepr, PConstanted, pconstantFromRepr, pconstantToRepr),
@@ -65,6 +70,44 @@ fixedNumerator (FixedDecimal num) = num
 -}
 fixedDenominator :: forall (exp :: Natural). (KnownNat exp) => FixedDecimal exp -> Integer
 fixedDenominator _ = 10 ^ natVal (Proxy @exp)
+
+{- | Exponent-changing multiplication (implemented with a single integer multiplication)
+
+ @since 3.11.0
+-}
+emul ::
+  forall (expA :: Natural) (expB :: Natural).
+  FixedDecimal expA ->
+  FixedDecimal expB ->
+  FixedDecimal (expA + expB)
+emul (FixedDecimal a) (FixedDecimal b) = FixedDecimal $ a * b
+
+{- | Exponent-changing division (implemented with a single integer division)
+
+ @since 3.11.0
+-}
+ediv ::
+  forall (expA :: Natural) (expB :: Natural).
+  FixedDecimal expA ->
+  FixedDecimal expB ->
+  FixedDecimal (expA - expB)
+ediv (FixedDecimal a) (FixedDecimal b) = FixedDecimal $ a `div` b
+
+{- | Convert to a different type-level exponent.
+
+ @since 3.11.0
+-}
+convertExp ::
+  forall (expA :: Natural) (expB :: Natural).
+  (KnownNat expA, KnownNat expB) =>
+  FixedDecimal expA ->
+  FixedDecimal expB
+convertExp (FixedDecimal a) =
+  let ediff = natVal (Proxy @expB) - natVal (Proxy @expA)
+   in FixedDecimal $
+        if ediff >= 0
+          then a * 10 ^ ediff
+          else a `div` 10 ^ (-ediff)
 
 instance (KnownNat exp) => Num (FixedDecimal exp) where
   (FixedDecimal a) + (FixedDecimal b) = FixedDecimal (a + b)
@@ -235,6 +278,28 @@ pfixedDenominator ::
   Term s PInteger
 pfixedDenominator _ = 10 ^ natVal (Proxy @unit)
 
+{- | Exponent-changing multiplication (implemented with a single integer multiplication)
+
+ @since 3.11.0
+-}
+pemul ::
+  forall (s :: S) (expA :: Natural) (expB :: Natural).
+  Term s (PFixedDecimal expA) ->
+  Term s (PFixedDecimal expB) ->
+  Term s (PFixedDecimal (expA + expB))
+pemul a b = pcon . PFixedDecimal $ pfixedNumerator a * pfixedNumerator b
+
+{- | Exponent-changing division (implemented with a single integer division)
+
+ @since 3.11.0
+-}
+pediv ::
+  forall (s :: S) (expA :: Natural) (expB :: Natural).
+  Term s (PFixedDecimal expA) ->
+  Term s (PFixedDecimal expB) ->
+  Term s (PFixedDecimal (expA - expB))
+pediv a b = pcon . PFixedDecimal $ pdiv # pfixedNumerator a # pfixedNumerator b
+
 {- | Change decimal point.
 
  *Caution* This function will drop precision when converting from more
@@ -256,9 +321,10 @@ pconvertExp = phoistAcyclic $
   plam $ \z ->
     let ediff = (natVal (Proxy @exp2) - natVal (Proxy @exp1))
      in pcon . PFixedDecimal $
-          if ediff > 0
-            then pto z * (10 ^ abs ediff)
-            else pdiv # pto z #$ 10 ^ abs ediff
+          case compare ediff 0 of
+            GT -> pto z * (10 ^ abs ediff)
+            EQ -> pto z
+            LT -> pdiv # pto z # (10 ^ (-ediff))
 
 {- | Convert 'PFixed' into 'PInteger'.